dbt_pumpkin/dbt_compat.py
from dbt.cli.main import dbtRunner, dbtRunnerResult, RunExecutionResult
from dbt.cli.resolvers import default_project_dir
from dbt.contracts.graph.manifest import Manifest
from dbt.contracts.graph.nodes import SourceDefinition, ModelNode, SnapshotNode, SeedNode

try:
    from dbt_common.events.base_types import EventMsg
except:
    from dbt.events.base_types import EventMsg

try:
    from dbt.artifacts.resources.v1.components import ColumnInfo
except:
    from dbt.contracts.graph.nodes import ColumnInfo
